fix compiler warning and relative path issue

+++ c++/BitBoard.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 const int directions[] = {1, 7, 6, 8};
 
-Bitboard::Bitboard(long *boards, int *heights, int *moves, int numberOfPiecesPlayed) {
+Bitboard::Bitboard(long long *boards, int *heights, int *moves, int numberOfPiecesPlayed) {
     this->boards = boards;
     this->heights = heights;
     this->moves = moves;
@@ -12,19 +12,19 @@ Bitboard::Bitboard(long *boards, int *heights, int *moves, int numberOfPiecesPla
 }
 
 void Bitboard::insertPiece(int columnIndex) {
-    long move = 1L << this->heights[columnIndex]++;
+    long long move = 1L << this->heights[columnIndex]++;
     this->boards[this->numberOfPiecesPlayed & 1] ^= move;
     this->moves[numberOfPiecesPlayed++] = columnIndex;
 }
 
 void Bitboard::undoMove() {
     int columnIndex = this->moves[--this->numberOfPiecesPlayed];
-    long move = 1L << --this->heights[columnIndex];
+    long long move = 1L << --this->heights[columnIndex];
     this->boards[this->numberOfPiecesPlayed & 1] ^= move;
 }
 
 bool Bitboard::isWin() {
-    long board = this->boards[(numberOfPiecesPlayed-1) & 1];
+    long long board = this->boards[(numberOfPiecesPlayed-1) & 1];
     for (int direction: directions) {
         if ((board & (board >> direction) & (board >> 2 * direction) & (board >> 3 * direction)) != 0) {
             return true;
@@ -33,11 +33,11 @@ bool Bitboard::isWin() {
     return false;
 }
 
-long Bitboard::getPositionCode() {
+long long Bitboard::getPositionCode() {
     return 2 * boards[0] + boards[1] + TOP_MASK;
 }
 
 bool Bitboard::canPlay(int column) {
-    long pretendedMove = 1L << this->heights[column];
+    long long pretendedMove = 1L << this->heights[column];
     return (TOP_MASK & pretendedMove) == 0;
 }
